Don't retry securitytrails ping API on HTTP 429
securitytrails.com's ping API returns 429 if you've exceeded the quota
for the API key. This delays the setup phase by 90 seconds. This fix
introduces a ping_retry_on_http_429 flag which can be set to False
to avoid this behaviour. The flag is set to False on the securitytrails
module.